refactor: k6 benchmarks, shorten code, better performance, k6 cloud
