+++ docker-compose.yml
@@ -114,10 +114,12 @@ services:
         soft: 50000
         hard: 50000
     environment:
-      - API_TOKEN
-      - API_HOST
-      - K6_PARALLEL_REQS_PER_USER
+      - K6_API_TOKEN
+      - K6_API_HOST
+      - K6_CLOUD_API_TOKEN
       - K6_SCENARIO
+      - K6_TABLE_MAPPING
+      - K6_PARALLEL_REQS_PER_USER
       - K6_CONST_VIRTUAL_USERS
       - K6_CONST_TOTAL_REQUESTS
       - K6_CONST_TIMEOUT
