+++ provisioning/stream/kubernetes/templates/benchmark/job.yaml
@@ -26,15 +26,17 @@ spec:
           image: docker.io/keboolabot/stream-benchmark:latest
           imagePullPolicy: Always
           command: ["k6"]
-          args: ["run", "--out", "output-statsd", "/scripts/k6/stream-api/static.js"]
+          args: ["run", "--no-thresholds", "--out", "output-statsd", "/scripts/k6/stream-api/main.js"]
           resources:
             requests:
               cpu: "1"
               memory: "4Gi"
           env:
-            - name: API_HOST
+            - name: DD_STATSD_PORT
+              value: "8125"
+            - name: K6_API_HOST
               value: "http://stream-api.stream.svc.cluster.local"
-            - name: API_TOKEN
+            - name: K6_API_TOKEN
               valueFrom:
                 secretKeyRef:
                   name: benchmark
@@ -43,8 +45,8 @@ spec:
               valueFrom:
                 fieldRef:
                   fieldPath: status.hostIP
-            - name: DD_STATSD_PORT
-              value: "8125"
+            - name: K6_OVERWRITE_SOURCE_HOST
+              value: "http://stream-http-source.stream.svc.cluster.local"
             - name: K6_STATSD_ADDR
               value: $(DD_AGENT_HOST):$(DD_STATSD_PORT)
             - name: K6_STATSD_NAMESPACE
@@ -52,13 +54,15 @@ spec:
             - name: K6_STATSD_PUSH_INTERVAL
               value: "5s"
             - name: K6_STATSD_BUFFER_SIZE
-              value: "1000"
+              value: "10"
             - name: K6_STATSD_ENABLE_TAGS
               value: "true"
-            - name: K6_PARALLEL_REQS_PER_USER
-              value: "10"
             - name: K6_SCENARIO
               value: "ramping"
+            - name: K6_TABLE_MAPPING
+              value: "static"
+            - name: K6_PARALLEL_REQS_PER_USER
+              value: "1"
             - name: K6_CONST_VIRTUAL_USERS
               value: "1000"
             - name: K6_CONST_TOTAL_REQUESTS
@@ -66,7 +70,7 @@ spec:
             - name: K6_CONST_TIMEOUT
               value: "20m"
             - name: K6_RAMPING_MAX_VIRTUAL_USERS
-              value: "100"
+              value: "1000"
             - name: K6_RAMPING_UP_DURATION
               value: "1m"
             - name: K6_RAMPING_STABLE_DURATION
